pyproject.toml
[project]
name = "crawl-to-markdown"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "marimo>=0.18.4",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
]
